ref(lib/abi)[mnt]: split in modules `Layout` stuff
